started breaking up server logic into importable module files. Added a cookies-js module to handle cookie parsing in the browser
