+++ client/browser.html
@@ -10,5 +10,6 @@
     <div id='nonceContainer'></div>
   </body>
   <script src="https://cdn.socket.io/socket.io-1.4.5.js"></script>
+  <script src="./lib/cookies.min.js"></script>
   <script src="./browser.js"></script>
 </html>
