+++ package.json
@@ -21,6 +21,7 @@
     "body-parser": "^1.15.2",
     "browser-request": "^0.3.3",
     "cookie-parser": "^1.4.3",
+    "cookies-js": "^1.2.2",
     "ejs": "^2.4.2",
     "express": "^4.14.0",
     "express-useragent": "^0.2.4",
